backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-chess
